Reports, random primary keys was sometimes causing an error with the new report system. Backported to: 15.1.24.0;15.2.3.0 Reviewed by: Jason *Changed PIn.Int to PIn.Long for definitions lookup
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@12264 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/ReportingComplex/ReportComplex.cs
@@ -368,8 +368,8 @@ namespace OpenDental.ReportingComplex {
 											if(newQuery.DictDefNames==null) {
 												return false;
 											}
-											if(newQuery.DictDefNames.ContainsKey(PIn.Int(newQuery.ReportTable.Rows[0][query.ColumnNameToSplitOn].ToString()))) {
-												displayText=newQuery.DictDefNames[PIn.Int(newQuery.ReportTable.Rows[0][query.ColumnNameToSplitOn].ToString())];
+											if(newQuery.DictDefNames.ContainsKey(PIn.Long(newQuery.ReportTable.Rows[0][query.ColumnNameToSplitOn].ToString()))) {
+												displayText=newQuery.DictDefNames[PIn.Long(newQuery.ReportTable.Rows[0][query.ColumnNameToSplitOn].ToString())];
 												newQuery.GetGroupTitle().Size=new Size((int)(grfx.MeasureString(displayText,newQuery.GetGroupTitle().Font).Width/grfx.DpiX*100+2),(int)(grfx.MeasureString(displayText,newQuery.GetGroupTitle().Font).Height/grfx.DpiY*100+2));
 												newQuery.GetGroupTitle().StaticText=displayText;
 											}
@@ -421,8 +421,8 @@ namespace OpenDental.ReportingComplex {
 								if(newQuery.DictDefNames==null) {
 									return false;
 								}
-								if(newQuery.DictDefNames.ContainsKey(PIn.Int(newQuery.ReportTable.Rows[0][query.ColumnNameToSplitOn].ToString()))) {
-									displayText=newQuery.DictDefNames[PIn.Int(newQuery.ReportTable.Rows[0][query.ColumnNameToSplitOn].ToString())];
+								if(newQuery.DictDefNames.ContainsKey(PIn.Long(newQuery.ReportTable.Rows[0][query.ColumnNameToSplitOn].ToString()))) {
+									displayText=newQuery.DictDefNames[PIn.Long(newQuery.ReportTable.Rows[0][query.ColumnNameToSplitOn].ToString())];
 									newQuery.GetGroupTitle().Size=new Size((int)(grfx.MeasureString(displayText,newQuery.GetGroupTitle().Font).Width/grfx.DpiX*100+2),(int)(grfx.MeasureString(displayText,newQuery.GetGroupTitle().Font).Height/grfx.DpiY*100+2));
 									newQuery.GetGroupTitle().StaticText=displayText;
 								}
